% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecrire_equations.R
\name{ecrire_eq}
\alias{ecrire_eq}
\title{Write R code associated to one of the equations of the Natura 3.0 growth simulator for prediction}
\usage{
ecrire_eq(vari, ess)
}
\arguments{
\item{vari}{Name of the variable to predict: n, st, v, hd or is}

\item{ess}{Group species code: sab, epn, epx, rt, ri, bop, peu, ft, or empty if the variable is hd or is}
}
\value{
A character string containing the equation for predicting the variable \code{vari} of group species \code{ess}
}
\description{
Write R code associated to one of the equations of the Natura 3.0 growth simulator for predicting : N, St or V for a species group, or HD or IS
}
\examples{
\dontrun{
ecrire_eq(vari="n", ess="bop")
}
}
