% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fichier_arbres_aveccov}
\alias{fichier_arbres_aveccov}
\title{Fichier d'intrant à l'échelle de l'arbre pour Natura, avec les covariables}
\format{
## `fichier_arbres_aveccov`
A data frame with 29 lignes et 31 colonnes:
\describe{
  \item{essence}{Code d'essence}
  \item{ID_PE}{Identifiant de la placette}
  \item{TIGE_HA}{Nombre de tiges dans la classe de DHP=dhpcm et de l'essence=essence}
  \item{dhpcm}{Classe de DHP (cm)}
  \item{ETAT}{Code d'état de l'arbre}
  \item{LATITUDE}{Latitude de la placette}
  \item{LONGITUDE}{Latitude de la placette}
  \item{SDOM_BIO}{Code du sous-domaine bioclimatique}
  \item{REG_ECO}{Code de la région écologique}
  \item{type_eco}{Code du type écologique}
  \item{ALTITUDE}{Altitude (m)}
  \item{ORIGINE}{Code de la perturbation d'origine}
  \item{temps}{Temps depuis l'orgine}
  \item{iqs_pot_EPN}{IQS potentiel de EPN}
  \item{iqs_pot_EPB}{IQS potentiel de EPB}
  \item{iqs_pot_BOP}{IQS potentiel de BOP}
  \item{iqs_pot_PIB}{IQS potentiel de PIB}
  \item{iqs_pot_PEX}{IQS potentiel de PEX}
  \item{iqs_pot_PIG}{IQS potentiel de PIG}
  \item{iqs_pot_SAB}{IQS potentiel de SAB}
  \item{iqs_pot_THO}{IQS potentiel de THO}
  \item{clay}{Pourcentage d'argile dans le 0-5cm}
  \item{silt}{Pourcentage de limon dans le 0-5cm}
  \item{sand}{Pourcentage de sable dans le 0-5cm}
  \item{cec}{Capacité d'échange cationique dans le 0-5cm}
  \item{oc}{Matière organique dans le 0-5cm}
  \item{ph}{pH dans le 0-5cm}
  \item{ptot}{Précipitation totales annuelles normale 30 ans}
  \item{t_ma}{Température annuelle moyenne  normale 30 ans}
  \item{prec_gs}{Précipitation durant la saison de croissance moyenne des 10 dernières années}
  \item{temp_gs}{Température durant la saison de croissance moyenne des 10 dernières années}
}
}
\usage{
fichier_arbres_aveccov
}
\description{
Deux placettes avec leur liste d'arbres
}
\examples{
fichier_arbres_aveccov
}
\keyword{datasets}
