% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fichier_arbres_etudes}
\alias{fichier_arbres_etudes}
\title{Fichier d'intrant des arbres-études pour Natura}
\format{
## `fichier_arbres_etudes`
A data frame with 6 lignes et 5 colonnes:
\describe{
  \item{ID_PE}{Identifiant de la placette}
  \item{ESSENCE}{Code d'essence}
  \item{etage}{Code d'étage de l'arbre}
  \item{dhpcm}{DHP de l'arbre (cm)}
  \item{hauteur}{Hauteur de l'arbre (m)}
}
}
\usage{
fichier_arbres_etudes
}
\description{
Deux placettes avec leur liste d'arbres-études
}
\examples{
fichier_arbres_etudes
}
\keyword{datasets}
