% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fichier_arbres_sanscov}
\alias{fichier_arbres_sanscov}
\title{Fichier d'intrant à l'échelle de l'arbres pour Natura, sans les covariables}
\format{
## `fichier_arbres_sanscov`
A data frame with 29 lignes et 13 colonnes:
\describe{
  \item{essence}{Code d'essence}
  \item{ID_PE}{Identifiant de la placette}
  \item{TIGE_HA}{Nombre de tiges dans la classe de DHP=dhpcm et de l'essence=essence}
  \item{dhpcm}{Classe de DHP (cm)}
  \item{ETAT}{Code d'état de l'arbre}
  \item{LATITUDE}{Latitude de la placette}
  \item{LONGITUDE}{Latitude de la placette}
  \item{SDOM_BIO}{Code du sous-domaine bioclimatique}
  \item{REG_ECO}{Code de la région écologique}
  \item{type_eco}{Code du type écologique}
  \item{ALTITUDE}{Altitude (m)}
  \item{ORIGINE}{Code de la perturbation d'origine}
  \item{temps}{Temps depuis l'orgine}
  \item{an_mes}{Année de la mesure}
}
}
\usage{
fichier_arbres_sanscov
}
\description{
Deux placettes avec leur liste d'arbres
}
\examples{
fichier_arbres_sanscov
}
\keyword{datasets}
