% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtrer_place.R
\name{Filtrer_place}
\alias{Filtrer_place}
\title{Sélectionne les placettes qui peuvent être simulées avec Natura 3.0}
\usage{
Filtrer_place(fichier)
}
\arguments{
\item{fichier}{Nom de la table à l'échelle de l'arbre ou de la placette avec les colonnes nécessaires}
}
\value{
Table dont les placettes ont été filtrées et contenant des nouvelles covariables.
}
\description{
Sélectionne les placettes qui peuvent être simulées avec Natura 3.0
}
\details{
Les placettes sont sélectionnées selon le type écologique (type_eco), le sous-domaine bioclimatique (sdom_bio) et la perturbation d'origine (origine).
Les variables végétation potentielle (veg_pot) et milieu physique (milieu) sont créées à partir de type_eco, des variables dichotomiques sont créées pour chaque niveau de sdom_bio,
veg_pot, milieu et origine. Les variables d'IQS potentiel sont renommées. La variable annee=0 est créée.
}
\examples{
\dontrun{
plots <- Filtrer_place(fichier=liste_plot)
}
}
