% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lecture_fichier_compile.R
\name{Lecture_compile}
\alias{Lecture_compile}
\title{Lire le fichier des placettes à simuler avec Natura et valider le nom des colonnes}
\usage{
Lecture_compile(file, iqs, climat, sol)
}
\arguments{
\item{file}{Nom du fichier des arbres à lire (table, Excel ou csv)}

\item{iqs}{Booléen
\itemize{
   \item \code{TRUE} (par défaut): si les IQS potentils doivent être extraits des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
   \item \code{FALSE}: si les IQS potentiels (m, hauteur dominante à 50 ans) sont fournis dans \code{file_arbre} ou \code{in file_compile}.
 }}

\item{climat}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables climatiques doivent être extraites des cartes. Les colonnes latitude, longitude et l'année de mesure de la placette doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \item \code{FALSE}: si les variables climatiques sont fournies dans \code{file_arbre} ou \code{file_compile}.
}}

\item{sol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables sur les propriétés des sols sont à extraires des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \item \code{FALSE}: si les variables sur les propriétés des sols sont fournies dans \code{file_arbre} ou dans \code{file_compile}.
}}
}
\value{
Table à l'échelle de la placette ou un message d'erreur s'il y a une erreur dans le nom des colonnes.
}
\description{
Lire le fichier des placettes à simuler avec Natura et valider le nom des colonnes.
}
