% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lecture_fichier_etudes.R
\name{Lecture_etudes}
\alias{Lecture_etudes}
\title{Function reading the study-tree file for a simulation with Natura 3.0}
\usage{
Lecture_etudes(file)
}
\arguments{
\item{file}{Name of the file to read (dataframe, Excel or csv file)}
}
\value{
Dataframe of filtered study-tree or an error message
}
\description{
Read a study-tree file for a simulation with Natura 3.0. Check the names of the variables and filter trees from variable dhp, hauteur and etage
}
