% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liste_essence}
\alias{liste_essence}
\title{Liste des essences acceptées}
\format{
## `liste_essence`
 Data frame avec la liste des essences acceptées et leur groupe
 \describe{
  \item{essence}{Code d'essence}
  \item{groupe_ess}{Groupe d'essences auquel appartient l'essence}
}
}
\usage{
liste_essence
}
\description{
Data frame de 2 colonnes et 34 lignes
}
\examples{
liste_essence
}
\keyword{datasets}
