% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liste_veg_pot}
\alias{liste_veg_pot}
\title{Liste des végétations potentielles acceptées}
\format{
## `liste_veg_pot`
 Vecteurs avec les 18 codes de végétations potentielles acceptées
}
\usage{
liste_veg_pot
}
\description{
Vecteurs de taille 18
}
\examples{
liste_veg_pot
}
\keyword{datasets}
