% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Natura_oneStep.R
\name{Natura_oneStep}
\alias{Natura_oneStep}
\title{Simulation of one time-step with Natura 3.0 growth model}
\usage{
Natura_oneStep(
  data,
  param,
  data_info,
  long_int,
  pas,
  dec_perturb = 0,
  dec_tbe1 = 0,
  tbe1 = 0,
  dec_tbe2 = 0,
  tbe2 = 0
)
}
\arguments{
\item{data}{Dataframe to be simulated. One line per plot. Must contain only the variables associated to stands characteristics to be simulated and plot id (no covariates an no paramteers)}

\item{param}{Dataframe (one line per plot) of parameter values of all the models of Natura 3.0 for each plot}

\item{data_info}{Dataframe (one line per plot) with only variables associated to fixed stand characteristics over time}

\item{long_int}{Lenght of simulation time step in years}

\item{pas}{The simulation step number}

\item{dec_perturb}{The simulation step number with a presence of perturbation (other than spruce budworm)}

\item{dec_tbe1}{The simulation step number with a presence of a first spruce budworm defoliation}

\item{tbe1}{Severity of the first spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}

\item{dec_tbe2}{The simulation step number with a presence of a second spruce budworm defoliation}

\item{tbe2}{Severity of the second spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}
}
\value{
Dataframe of predicted stand characteristics
}
\description{
Simulation of one time-step with Natura 3.0 growth model, usually of 10 years, of the growth of characteristics of a stand (number of trees, basal area and volume per group species, dominant height and shannon index), based on plot origine and time since origine.
}
