% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametres_stochastiques.R
\name{param_evol_n_st_v_stoch}
\alias{param_evol_n_st_v_stoch}
\title{Génère les paramètres des modèles d'évolution de Nxxx, STxxx et Vxxx}
\usage{
param_evol_n_st_v_stoch(
  liste_place,
  mode_simul = "DET",
  nb_iter = 1,
  horizon,
  liste_ess,
  seed_value = NULL
)
}
\arguments{
\item{liste_place}{Vecteur contenant les identifiants des placettes}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{nb_iter}{Si mode stochastique, le nombre d'itérations à effectuer (> 1), ignoré si \code{mode_simul="DET"}}

\item{horizon}{Nombre de décennies à simuler, un chiffre de 1 à 15.}

\item{liste_ess}{Vecteur contenant les codes des groupes d'essences de Natura}

\item{seed_value}{Optionnel. Valeur de départ pour la génération de nombres aléatoires. Utilisé lors des tests pour la reproductibilité des résultats.}
}
\value{
Une liste de listes. Une liste par groupe d'essences, et pour chaque groupe d'essences, une liste de 3 éléments:
\enumerate{
  \item essence: chaine de caractères contenant le code du groupe d'essences
  \item effet_fixe: table contenant les paramètres des effets fixes, une ligne par pas de simulation et itération
  \item erreur_residuelle: table contenant les valeurs des erreurs résiduelles, une ligne par placette, pas de simulation et itération
}
}
\description{
Génère les paramètres des modèles d'évolution de Nxxx, STxxx et Vxxx (où xxx est chacun des 8 groupes d'essences, donc 24 modèles), déterministe ou stochastique, pour chacune des placettes pour tous les pas de simulation et toutes les itérations
}
\details{
Les modèles de Nxxx, STxxx et Vxxx forment un système d'équations linéaires, calibrés séparément par groupe d'essences. Il n'y a pas d'effets aléatoires.
Les erreurs résiduelles sont corrélées entre les 3 équations. Il y a une matrice de covariances des effets fixes.
}
