% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametres_stochastiques.R
\name{param_hdom0_stoch}
\alias{param_hdom0_stoch}
\title{Generating parameters of the global tree height model for step 0 domimant height estimation}
\usage{
param_hdom0_stoch(liste_arbre, nb_iter, mode_simul)
}
\arguments{
\item{liste_arbre}{Dataframe with tree ID (no_arbre) and plot Id (id_pe) for which height must be predicted for dominant height estimation}

\item{nb_iter}{Number of iteration if stochastic mode, ignored if \code{mode_simul="DET"}}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}
}
\value{
A list of 2 elements:
\enumerate{
  \item effet_fixe: dataframe of fixed effect parameter values (one line per iteration)
  \item erreur_arbre: dataframe of residual error values (random tree effect values and residual error values, one line per tree per iteration)
}
}
\description{
Generate parameters of the global tree height model for step 0 dominant height estimation, deterministic or stochastic, for each tree of each plot and for all iterations.
}
\details{
The global tree height estimation model for dominant height estimation is a non linear mixed model calibrated globally for all species. There is a tree random effect
and the residual errors are not corralated. There is no covariance matrix of fixed effects, since there is only one parameter.
This model is used when a species is present in the tree list but not in the study-tree list.
}
