% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametres_stochastiques.R
\name{param_is_evol_stoch}
\alias{param_is_evol_stoch}
\title{Génère les paramètres du modèle d'évolution de l'indice de Shannon}
\usage{
param_is_evol_stoch(
  liste_place,
  mode_simul = "DET",
  nb_iter = 1,
  horizon,
  seed_value = NULL
)
}
\arguments{
\item{liste_place}{Vecteur contenant les identifiants des placettes.}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{nb_iter}{Si mode stochastique, le nombre d'itérations à effectuer (> 1), ignoré si \code{mode_simul="DET"}}

\item{horizon}{Nombre de décennies à simuler, un chiffre de 1 à 15.}

\item{seed_value}{Optionnel. Valeur de départ pour la génération de nombres aléatoires. Utilisé lors des tests pour la reproductibilité des résultats.}
}
\value{
Une liste de listes. Une liste par groupe d'essences, et pour chaque groupe d'essences, une liste de 4 éléments:
\enumerate{
  \item essence: chaine de caractère contenant le code du groupe d'essences
  \item effet_fixe: table contenant les paramètres des effets fixes, une ligne par pas de simulation et itération
  \item random_placette: table contenant les valeurs des effets aléatoires, une ligne par placette, pas de simulation et itération
  \item erreur_residuelle: table contenant les valeurs des erreurs résiduelles, une ligne par placette, pas de simulation et itération
}
}
\description{
Génère les paramètres du modèle d'évolution de l'indice de Shannon, déterministe ou stochastique, pour chacune des placettes et tous leurs pas de simulation, pour toutes les itérations.
}
\details{
Le modèle est un modèle linéaire mixte calibré séparément par groupe d'essences dominantes, avec un effet aléatoire de placette.
Il n'y a pas de matrice de corrélation sur les résidus. Il y a une matrice de covariance des effets fixes.
}
