% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_simul.R
\name{plot_simul}
\alias{plot_simul}
\title{Produit un graphique du résultat d'une simulation déterministe}
\usage{
plot_simul(data, variable, orig = TRUE)
}
\arguments{
\item{data}{Table contenant les résultats d'une simulation déterministe avec la fonction \code{SimulNatura}}

\item{variable}{Le nom de la variable pour laquelle on veut un graphique, ex: sttot, stepn, hd, etc.}

\item{orig}{TRUE: facet_wrap(~origine) (par défaut), FALSE: sans facet-wrap}
}
\value{
Un graphique
}
\description{
\code{plot_simul()} produit un graphique du résultat d'une simulation déterministe pour une variable.
}
\examples{
\dontrun{
data_simul <- SimulNatura(file_arbre=fichier_arbres_sanscov,
file_etude=fichier_arbres_etudes, horizon=5)
plot_simul(data=data_simul, variable=sttot)
}

}
