% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_simul.R
\name{plot_simul_ess}
\alias{plot_simul_ess}
\title{Produit un graphique par essence du résultat d'une simulation déterministe}
\usage{
plot_simul_ess(data, variable, mode_simul = "DET")
}
\arguments{
\item{data}{Table contenant les résultats d'une simulation déterministe avec la fonction \code{SimulNatura}}

\item{variable}{Le nom de la variable pour laquelle on veut un graphique: 'st', 'v', 'n', ou 'dq'}
}
\value{
Un graphique par essence
}
\description{
\code{plot_simul_ess()} produit un graphique par essence du résultat d'une simulation déterministe pour une variable.
}
\examples{
\dontrun{
data_simul <- SimulNatura(file_arbre=fichier_arbres_sanscov, file_etude=fichier_arbres_etudes, horizon=5)
plot_simul_ess(data=data_simul, variable='st')
}

}
