% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_simul.R
\name{plot_simul_sto}
\alias{plot_simul_sto}
\title{Produit un graphique du résultat d'une simulation stochastique}
\usage{
plot_simul_sto(data, variable)
}
\arguments{
\item{data}{Table contenant les résultats d'une simulation stochastique avec la fonction \code{SimulNatura}}

\item{variable}{Le nom de la variable pour laquelle on veut un graphique, ex: sttot, vepn, hd}
}
\value{
Un graphique
}
\description{
\code{plot_simul_sto()} produit un graphique du résultat d'une simulation stochastique pour une variable.
}
\details{
un graphique
}
\examples{
\dontrun{
data_simul_sto <- SimulNatura(file_arbre=fichier_arbres_sanscov,
file_etude=fichier_arbres_etudes, horizon=5, mode_simil='STO', nb_iter=30)
plot_simul_sto(data=data_simul_sto, variable=sttot)
}

}
