% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_fichier_arbres.R
\name{Prep_arbres}
\alias{Prep_arbres}
\title{Plot compilation from tree list}
\usage{
Prep_arbres(
  fic_arbre,
  ht,
  vol,
  iteration,
  parametre_ht_dhp,
  parametre_vol,
  mode_simul,
  liste_ess_ht
)
}
\arguments{
\item{fic_arbre}{Dataframe with one line per tree per plot, with variables : id_pe, sdom_bio, altitude, veg_pot, dummy vp, milieu, dummy mp, type_eco, p_tot, t_ma, prec_gs, temp_gs, dummy orig, temps, all iqs, ph, cec, oc, sand, clay, silt, no_arbre, essence, dhpcm, nb_tige, tige_ha, etat, hauteur_pred, vol_dm3}

\item{ht}{A boolean
\itemize{
  \item \code{TRUE} (default): if tree height must be estimated, when \code{file_arbre} is specified
  \item \code{FALSE}: if tree height is in \code{file_arbre}, ignored if \code{file_compile} is specified. Column name must be hauteur_pred (in meters)
  }}

\item{vol}{A boolean
\itemize{
  \item \code{TRUE} (default): if tree volume must be estimated, when \code{file_arbre} is specified
  \item \code{FALSE}: if tree volume is in \code{file_arbre}, ignored if \code{file_compile} is specified. If provided, column name must be vol_dm3 (in dm3)
  }}

\item{iteration}{Iteration number}

\item{parametre_ht_dhp}{Object with the tree height model parameter values}

\item{parametre_vol}{Object with the tree volume model parameter values}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{liste_ess_ht}{Vector of species code name with a tree height model}
}
\value{
A list of 2 elements: dataframe of tree characteristics (with height and volume) and dataframe of plot characteristics ready to be simulated
}
\description{
Estimate tree height and/or tree volume if necessary, calculate shanon index from trees dhp, calculate stand characterics (n-st-v) per species group
}
