% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_fichier_compile.R
\name{Prep_compile}
\alias{Prep_compile}
\title{Préparation du fichier compilé à la step 0 pour la simulation dans Natura}
\usage{
Prep_compile(fichier_compile)
}
\arguments{
\item{fichier_compile}{Table avec une ligne par placette}
}
\value{
Le fichier \code{fichier_compile} avec toutes les variables nécessaires au modèle Natura
}
\description{
Préparation du fichier compilé à la step 0 pour la simulation dans Natura: création des variables nécessaires et retrait des variables inutiles
}
