% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_fichier_etude.R
\name{Prep_etude}
\alias{Prep_etude}
\title{Calculation of dominant height at step 0 from study trees and top trees for a Natura 3.0 simulation}
\usage{
Prep_etude(
  fic_etude,
  fic_arbre,
  iteration,
  mode_simul,
  param_hdom0_ess,
  param_hdom0_global
)
}
\arguments{
\item{fic_etude}{Dataframe of the study-trees with mesured height}

\item{fic_arbre}{Dataframe of the trees, with estimated height}

\item{iteration}{Iteration number}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{param_hdom0_ess}{Object with the species height model parameter values, model for step 0 dominant height estimation}

\item{param_hdom0_global}{Object with the global height model parameter values, model for step 0 dominant height estimation}
}
\value{
Dataframe, one line per plot, with step 0 estimated plot domimant height
}
\description{
Calculation of dominant height at step 0 from study trees and top trees for a Natura 3.0 simulation.
}
