% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_parametre.R
\name{Prep_parametre}
\alias{Prep_parametre}
\title{Select the growth model parameters associated with a time step and an iteration for a list of plots for a simulation with Natura 3.0}
\usage{
Prep_parametre(
  data,
  data_info,
  iteration,
  pas,
  mode_simul,
  param_is_evol,
  param_hd_evol,
  param_n_st_v,
  liste_ess
)
}
\arguments{
\item{data}{Dataframe with only plot id (id_pe) and variables of percentage of group species basal area (pct_epn, pct_epx, etc)}

\item{data_info}{Dataframe with only fixed stand characteristics over time}

\item{iteration}{Iteration number}

\item{pas}{Time step number}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{param_is_evol}{Object with shannon index growth model parameter values for all time steps and all iterations}

\item{param_hd_evol}{Object with dominant height growth model parameter values for all time steps and all iterations}

\item{param_n_st_v}{Object with n-st-v growth models parameter values for all time steps and all iterations}

\item{liste_ess}{Vector with Natura group species code names}
}
\value{
Dataframe of plot characteristics and all their growth models parameter values for a time step and an iteration, ready to be simulated
}
\description{
Select and add to plot dataframe all the growth models parameters associated with a time step and an iteration.
It also determine the dominant species of the plot and select iqs associated with this species.
}
