% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simul_oneIter.R
\name{simul_oneIter_arbre}
\alias{simul_oneIter_arbre}
\title{Complete simulation (all time steps) of one iteration with Natura 3.0 simulator when the start file is at the tree scale}
\usage{
simul_oneIter_arbre(
  fic_arbre,
  fic_etude,
  ht,
  vol,
  horizon,
  iteration,
  mode_simul,
  param_is_evol,
  param_hdom0_ess,
  param_hdom0_global,
  parametre_ht_dhp,
  parametre_vol,
  liste_ess_ht,
  param_hd_evol,
  param_n_st_v,
  liste_ess,
  long_int,
  dec_perturb,
  dec_tbe1,
  tbe1,
  dec_tbe2,
  tbe2
)
}
\arguments{
\item{fic_arbre}{Dataframe of trees per plots, filtered and with the covariates}

\item{fic_etude}{Dataframe of study trees per plots}

\item{ht}{A boolean
\itemize{
  \item \code{TRUE} (default): if tree height must be estimated, when \code{file_arbre} is specified
  \item \code{FALSE}: if tree height is in \code{file_arbre}, ignored if \code{file_compile} is specified. Column name must be hauteur_pred (in meters)
  }}

\item{vol}{A boolean
\itemize{
  \item \code{TRUE} (default): if tree volume must be estimated, when \code{file_arbre} is specified
  \item \code{FALSE}: if tree volume is in \code{file_arbre}, ignored if \code{file_compile} is specified. If provided, column name must be vol_dm3 (in dm3)
  }}

\item{horizon}{Number of 10-year time steps to simulate, 1 à 15}

\item{iteration}{Iteration number}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{param_is_evol}{Object with shannon index growth model parameter values for all time steps and all iterations}

\item{param_hdom0_ess}{Object with step 0 species tree height estimation model parameter values for all iterations (for HD estimation at step 0)}

\item{param_hdom0_global}{Object with step 0 global tree height estimation model parameter values for all iterations (for HD estimation at step 0)}

\item{parametre_ht_dhp}{Object with tree height estimation model parameter values for all time steps and all iterations}

\item{parametre_vol}{Object with tree volume estimation model parameter values for all time steps and all iterations}

\item{liste_ess_ht}{Vector with species code name for which there is a tree height model}

\item{param_hd_evol}{Object with dominant height growth model parameter values for all time steps and all iterations}

\item{param_n_st_v}{Object with n-st-v growth models parameter values for all time steps and all iterations}

\item{liste_ess}{Vector with Natura group species code names}

\item{long_int}{Length of time steps (years)}

\item{dec_perturb}{The simulation step number with a presence of perturbation (other than spruce budworm)}

\item{dec_tbe1}{The simulation step number with a presence of a first spruce budworm defoliation}

\item{tbe1}{Severity of the first spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}

\item{dec_tbe2}{The simulation step number with a presence of a second spruce budworm defoliation}

\item{tbe2}{Severity of the second spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}
}
\value{
Dataframe with one line per plot per time step with predicted values of HD and IS, and N, ST and V per group species
}
\description{
Do a complete simulation (all time steps) of one iteration with Natura 3.0 simulator when the start file is at the tree scale
}
