% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simul_oneIter.R
\name{simul_oneIter_compile}
\alias{simul_oneIter_compile}
\title{Complete simulation (all time steps) for one iteration with Natura 3.0 simulator when the start file is already compiled at the plot scale}
\usage{
simul_oneIter_compile(
  fichier,
  horizon,
  iteration,
  PlacT0,
  data_info,
  mode_simul,
  param_is_evol,
  param_hd_evol,
  param_n_st_v,
  liste_ess,
  long_int,
  dec_perturb,
  dec_tbe1,
  tbe1,
  dec_tbe2,
  tbe2
)
}
\arguments{
\item{fichier}{Dataframe with stand characteristics at step 0 and all covariates, ready to be simulated, one line per plot}

\item{horizon}{Number of 10-year time steps to simulate, 1 à 15}

\item{iteration}{Iteration number}

\item{PlacT0}{Dataframe with only stand characteristics at step 0 (no covariates)}

\item{data_info}{Dataframe with only fixed stand characteristics over time}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{param_is_evol}{Object with shannon index growth model parameter values for all time steps and all iterations}

\item{param_hd_evol}{Object with dominant height growth model parameter values for all time steps and all iterations}

\item{param_n_st_v}{Object with n-st-v growth models parameter values for all time steps and all iterations}

\item{liste_ess}{Vector with Natura group species code names}

\item{long_int}{Length of time steps (years)}

\item{dec_perturb}{The simulation step number with a presence of perturbation (other than spruce budworm)}

\item{dec_tbe1}{The simulation step number with a presence of a first spruce budworm defoliation}

\item{tbe1}{Severity of the first spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}

\item{dec_tbe2}{The simulation step number with a presence of a second spruce budworm defoliation}

\item{tbe2}{Severity of the second spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}
}
\value{
Dataframe with one line per plot per time step with predicted values of HD and IS, and N, ST and V per group species
}
\description{
Do a Complete simulation (all time steps) for one iteration with Natura 3.0 simulator when the start file is already compiled at the plot scale.
}
