% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simul_oneIter.R
\name{simul_oneIter_compileV2}
\alias{simul_oneIter_compileV2}
\title{Effectue une simulation complète (tous les pas de simulation) pour une itération, lorsque le fichier en entrée est à l'échelle de la placette, version à utiliser lorsque les paramètres IS et HS sont dans le même objet}
\usage{
simul_oneIter_compileV2(
  fichier,
  horizon,
  iteration,
  PlacT0,
  data_info,
  mode_simul,
  param_ishd_evol,
  param_n_st_v,
  liste_ess,
  long_int,
  dec_perturb,
  dec_tbe1,
  tbe1,
  dec_tbe2,
  tbe2
)
}
\arguments{
\item{fichier}{Table contenant toutes les variables nécessaires pour effectuer une simulation, incluant les paramètres, une ligne par placette}

\item{horizon}{Nombre de décennies à simuler, un chiffre de 1 à 15.}

\item{iteration}{Numéro de l'itération}

\item{PlacT0}{Table contenant seulement les colonnes associées aux caractéristiques dendrométriques des placettes}

\item{data_info}{Table contenant seulement les colonnes associées aux caractéristiques des placettes qui ne changent pas dans le temps}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{param_ishd_evol}{Objet contenant les paramètres du modèle de l'indice de Shannon et du modèle de hauteur dominante, pour toutes les itérations et pas de simulation}

\item{param_n_st_v}{Objet contenant les paramètres des modèles de Nxxx, STxxx et Vxxx, pour toutes les itération et pas de smulation}

\item{liste_ess}{Vecteur contenant les codes des groupes d'essences de Natura}

\item{long_int}{Nombre d'années correspondant à un pas de simulation}

\item{dec_perturb}{Optionnel. Le numéro de la décennie dans laquelle on simule la présence d'une perturbation partielle autre qu'une épidémie de TBE (chablis partiel, verglas pertiel, dépérissement partiel, brulis partiel).}

\item{dec_tbe1}{Optionnel. Le numéro de la décennie dans laquelle on simule une défoliation due à la TBE. \code{tbe1} doit alors est spécifié.}

\item{tbe1}{Optionnel. Indice de réduction de croissance due à la TBE durant la décennie \code{dec_tbe1}, un entier 1 à 5. \code{dec_tbe1} doit alors est spécifié.
\itemize{
  \item 1 = 1 année de défoliation sévères
  \item 2 = 4 années de défoliations sévères
  \item 3 = 6 années de défoliations sévères
  \item 4 = 8 années de défoliations sévères
  \item 5= 10 années de défoliations sévères.
  }}

\item{dec_tbe2}{Optionnel. Le numéro de la 2e décennie dans laquelle on simule une défoliation due à la TBE. \code{tbe2} doit alors est spécifié.}

\item{tbe2}{Optionnel. Indice de réduction de croissance due à la TBE durant la décennie \code{dec_tbe2}, un entier 1 à 5. Voir \code{tbe1} pour définition des niveaux.}
}
\value{
Table avec une ligne par placette et pas de simulation contenant les prévision de HD et IS et N, ST et V de chacun des groupes d'essences
}
\description{
Effectue une simulation complète (tous les pas de simulation) pour une itération, lorsque le fichier en entrée est à l'échelle de la placette, version à utiliser lorsque les paramètres IS et HS sont dans le même objet.
}
