% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateurNatura.R
\name{SimulNatura}
\alias{SimulNatura}
\title{Main function of Natura 3.0 growth simulator}
\usage{
SimulNatura(
  file_arbre,
  file_etude,
  file_compile,
  file_export,
  horizon,
  mode_simul = "DET",
  nb_iter = 1,
  iqs = TRUE,
  climat = TRUE,
  sol = TRUE,
  ht = TRUE,
  vol = TRUE,
  dec_perturb = 0,
  dec_tbe1 = 0,
  tbe1 = 0,
  dec_tbe2 = 0,
  tbe2 = 0
)
}
\arguments{
\item{file_arbre}{Name of the file containing tree and plot informations (dataframe, xlsx, csv). One line per tree or per species/dbh class.
If not specified, \code{file_compile} must be provided. Column names must be:
 \itemize{
   \item id_pe: identifiant unique de la placette
   \item origine: code d'origine de la placette (ES: épidémie sévère, BR: brulis, CT: coupe totale)
   \item temps_origine: temps depuis l'origine (ans)
   \item sdom_bio: sous-domaine bioclimatique, en majuscule (ex: 2E, 4O)
   \item altitude: altitude en m
   \item type_eco: type écologique (ex: MS22)
   \item essence: code d'essence de l'arbre (ex: SAB, EPN, BOP)
   \item dhpcm: dhp (cm) de l'arbre ou classe de dhp (seul les arbres de plus de 9 cm sont retenus)
   \item nb_tige: nombre d'arbres de l'essence et de la classe de dhp, à l'ha
   \item etat: code d'état de l'arbre, seuls les vivants sont retenus, codes 10, 12, 30, 32, 40, 42, 50, 52
 }}

\item{file_etude}{Name of the file containing the study trees (dataframe, xlsx, csv). One line per tree.
if not specified, \code{file_compile} must be specified. Column names must be:
\itemize{
  \item id_pe: identifiant unique de la placette
  \item étage: code d'étage de l'arbre: C, D, O, I, V, seul les C-D sont retenus
  \item essence: code d'essence de l'arbre (ex: SAB, EPN, BOP)
  \item dhpcm: dhp de l'arbre-étude (cm)
  \item hauteur: hauteur de l'arbre étude (m)
}}

\item{file_compile}{Name of the file containig the compiled plots (dataframe, xlsx, csv). One line per plot.
 if not specified, \code{file_arbre} and \code{file_etude} must be provided. Column names must be:
 \itemize{
   \item id_pe: identifiant unique de la placette
   \item origine: code d'origine de la placette (ES: épidémie sévère, BR: brulis, CT: coupe totale)
   \item temps_origine: temps depuis l'origine (ans)
   \item sdom_bio: sous-domaine bioclimatique, en majuscule (ex: 2E, 4O)
   \item altitude: altitude en m
   \item type_eco: type écologique (ex: MS22)
   \item is: indice de structure diamètrale de Shannon (valeur entre 0 et 1; 0 : toutes les tiges sont dans la même classe de dhp; 1 : les tiges sont également distribuées dans 19 classes de dhp)
   \item hd: hauteur dominante, hauteur moyenne des 100 plus gros arbres à l'ha (m)
   \item nxxx : nombre d'arbres de plus de 9cm du groupe d'essences xxx (/ha), mettre 0 si absent. xxx : each of 8 species groups (bop, peu, ft, ri, rt, epx, epn, sab)
   \item stxxx: surface terrière marchande du groupe d'essences xxx (m2/ha), mettre 0 si absent. xxx : each of 8 species groups (bop, peu, ft, ri, rt, epx, epn, sab)
   \item vxxx: volume marchand brut du groupe d'essences xxx (m3/ha), mettre 0 si absent. Si aucun volume, mettre les 8 groupes à 0 et l'évolution du volume ne sera pas effectué. xxx : each of 8 species groups (bop, peu, ft, ri, rt, epx, epn, sab)
}}

\item{file_export}{Name of the file to export, will contain the simulation results (format csv), facultative}

\item{horizon}{Number of 10-year time steps to simulate, 1 à 15}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{nb_iter}{Number of iteration if stochastic mode, ignored if \code{mode_simul="DET"}}

\item{iqs}{A boolean
\itemize{
   \item \code{TRUE} (default): if potential site index are to be extrated from maps. Columns latitude and longitude must be provided in \code{file_arbre} or in \code{file_compile}.
   \itemize{
      \item latitude, longitude: coordinates of plots, in decimal degrees
   }
   \item \code{FALSE}: if potential site index (m) are provided in \code{file_arbre} or \code{in file_compile}. Column names must be:
   \itemize{
     \item iqs_pot_sab, iqs_pot_epn, iqs_pot_epb, iqs_pot_pig, iqs_pot_tho, iqs_pot_pib, iqs_pot_bop, iqs_pot_pex.
   }
 }}

\item{climat}{A boolean
\itemize{
   \item \code{TRUE} (default): if climate variables are to be extrated from maps. Columns latitude, longitude, an_mes must be provided in \code{file_arbre} or in \code{file_compile}.
   \itemize{
     \item latitude, longitude: coordinates of plots, in decimal degrees
     \item an_mes: measurement year of the plot
   }
  \item \code{FALSE}: if climate variables are in \code{file_arbre} or in \code{file_compile}. Column names must be:
  \itemize{
    \item p_tot: 1980-2010 mean total precipitations (mm)
    \item t_ma: 1980-2010 mean annual temprature (Celcius)
    \item prec_gs: last 10-years mean growing season precipitations (mm)
    \item temp_gs: last 10-years mean growing season temperature (Celcius)
  }
}}

\item{sol}{A boolean
\itemize{
  \item \code{TRUE} (default): if soil variables are to be extrated from maps. Columns latitude and longitude must be provided in \code{file_arbre} or in \code{file_compile}.
  \itemize{
    \item latitude, longitude: coordinates of plots, in decimal degrees
    }
  \item \code{FALSE}: if SIIGSOL soil variables are in \code{file_arbre} or in \code{file_compile}. Soils properties at depth 0-5 cm. Column names must be:
  \itemize{
    \item ph: pH
    \item clay: fraction of clay
    \item silt: fraction of silt
    \item sand: fraction of sand
    \item cec: capacité d'échange cationique
    \item oc: matière organique
  }
}}

\item{ht}{A boolean
\itemize{
  \item \code{TRUE} (default): if tree height must be estimated, when \code{file_arbre} is specified
  \item \code{FALSE}: if tree height is in \code{file_arbre}, ignored if \code{file_compile} is specified. Column name must be hauteur_pred (in meters)
  }}

\item{vol}{A boolean
\itemize{
  \item \code{TRUE} (default): if tree volume must be estimated, when \code{file_arbre} is specified
  \item \code{FALSE}: if tree volume is in \code{file_arbre}, ignored if \code{file_compile} is specified. If provided, column name must be vol_dm3 (in dm3)
  }}

\item{dec_perturb}{The simulation step number with a presence of perturbation (other than spruce budworm)}

\item{dec_tbe1}{The simulation step number with a presence of a first spruce budworm defoliation}

\item{tbe1}{Severity of the first spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}

\item{dec_tbe2}{The simulation step number with a presence of a second spruce budworm defoliation}

\item{tbe2}{Severity of the second spruce budworm defoliation, integer 0 to 5, 0  : no defoliation, 5 : 100 \% defoliation}
}
\value{
A data frame with the simulation results (HD, IS, Nxxx, STxxx, Vxxx, DQxxx), one line per plot/time step/iteration
}
\description{
\code{SimulNatura()} is the main function of Natura 3.0 growth simulator, the version using plot origine and time since origine.
}
\examples{
\dontrun{
# Simulation déterministe sur 50 ans à partir d'un fichier à l'échelle de l'arbre, en ne fournissant pas les covariables nécessaires
data_simul <- SimulNatura(file_arbre=fichier_arbres_aveccov, file_etude=fichier_arbres_etudes, horizon=5)

# Simulation stochastique sur 50 ans à partir d'un fichier à l'échelle de l'arbre, en fournissant les covariables nécessaires
data_simul <- SimulNatura(file_arbre=fichier_arbres_aveccov, file_etude=fichier_arbres_etudes, horizon=5, mode_simil='STO', iqs=FALSE, sol=FALSE, climat=FALSE)

# Simulation déterministe sur 50 ans à partir d'un fichier à l'échelle de la placette, en ne fournissant pas les covariables nécessaires
data_simul <- SimulNatura(file_compile=fichier_compile_sanscov, horizon=5)
}
}
