% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_simul.R
\name{sommaire_sto}
\alias{sommaire_sto}
\title{Calcul du sommaire des itérations en mode stochastique}
\usage{
sommaire_sto(data, variable)
}
\arguments{
\item{data}{Table contenant les résultats d'une simulation en mode stochastique avec la fonction \code{SimulNatura}}

\item{variable}{Le nom de la variable pour laquelle on veut un sommaire des itérations}
}
\value{
Une table avec la moyenne, le 2.5 et le 97.5 percentile
}
\description{
\code{sommaire_sto()} calcule le sommaire des itérations en mode stochastique.
}
\details{
Calcule de la moyenne, du 97.5 et 2.5 percentile des itérations par placette.
}
\examples{
\dontrun{
data_simul_sto <- SimulNatura(file_arbre=fichier_arbres_sanscov, file_etude=fichier_arbres_etudes, horizon=5, mode_simil='STO', nb_iter=30)
sommaire_sto(data=data_simul_sto, variable=sttot)
}
}
