% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_arbres.R
\name{valid_arbre}
\alias{valid_arbre}
\title{Vérifier le contenu des variables des fichier d'entrée}
\usage{
valid_arbre(type_fic, fichier)
}
\arguments{
\item{type_fic}{Type du fichier à vérifier: arbres, etudes}

\item{fichier}{Nom de la table à vérifier}
}
\value{
Table ou message d'erreur
}
\description{
Vérifier le contenu des variables des fichiers d'entrée
}
