% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validation_placettes.R
\name{valid_placette}
\alias{valid_placette}
\title{Vérifier le contenu des variables des fichier d'entrée}
\usage{
valid_placette(
  type_fic,
  fichier,
  ht = NULL,
  iqs = NULL,
  climat = NULL,
  sol = NULL
)
}
\arguments{
\item{type_fic}{Type du fichier à vérifier: arbres ou compil ou valid}

\item{fichier}{Nom de la table à vérifier}

\item{ht}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si la hauteur des arbres doit être estimée, seulement pour un fichier \code{file_arbre}
  \item \code{FALSE}: si la hauteur des arbres est fournie dans \code{file_arbre}, ignoré si \code{file_compile}. Le nom de la colonne doit être hauteur_pred (en mètres)
  }}

\item{iqs}{Booléen
\itemize{
   \item \code{TRUE} (par défaut): si les IQS potentils doivent être extraits des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
   \item \code{FALSE}: si les IQS potentiels (m, hauteur dominante à 50 ans) sont fournis dans \code{file_arbre} ou \code{in file_compile}.
 }}

\item{climat}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables climatiques doivent être extraites des cartes. Les colonnes latitude, longitude et l'année de mesure de la placette doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \item \code{FALSE}: si les variables climatiques sont fournies dans \code{file_arbre} ou \code{file_compile}.
}}

\item{sol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables sur les propriétés des sols sont à extraires des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \item \code{FALSE}: si les variables sur les propriétés des sols sont fournies dans \code{file_arbre} ou dans \code{file_compile}.
}}
}
\value{
Table ou message d'erreur
}
\description{
Vérifier le contenu des variables des fichiers d'entrée
}
