% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_columns.R
\name{valide_placette_etudes}
\alias{valide_placette_etudes}
\title{Valide la correspondance des placettes entre le fichier des arbres et celui des arbres-études}
\usage{
valide_placette_etudes(data_arbre, data_etude)
}
\arguments{
\item{data_arbre}{Nom du fichier contenant les informations sur les arbres et les placettes}

\item{data_etude}{Nom du fichier contenant les arbres-études}
}
\value{
Une liste de messages d'erreurs. Si aucune erreur n'est trouvée, une liste vide est retournée.
}
\description{
Cette fonction vérifie qu'au moins une des placettes présentes dans le fichier des arbres est également présente dans le fichier des arbres-études
}
