% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecrire_equations.R
\name{ecrire_eq}
\alias{ecrire_eq}
\title{Génère en code R les équations du simulateur de croissance Natura 3.0}
\usage{
ecrire_eq(vari, ess)
}
\arguments{
\item{vari}{Nom de la variable réponse pour laquelle on veut l'équation : n, st, v, hd ou is}

\item{ess}{Code du groupe d'essences pour lequel on veut l'équation, si \code{vari} est n, st ou v : sab, epn, epx, rt, ri, bop, peu, ft. Vide pour \code{vari=is}. Si \code{vari=hd},
utiliser ce paramètre pour générer chacune des parties de l'équation: a pour générer la partie associé au paramètre a, c pour générer la partie associée au paramètre c,
vide pour générer la forme générale de l'équation hd}
}
\value{
Une chaine de caractères contenant l'équation pour prédire la variable \code{vari} du groupe d'essences \code{ess}
}
\description{
Génère en code R les équations du simulateur de croissance Natura 3.0
}
\examples{
\dontrun{
ecrire_eq(vari="n", ess="bop")
ecrire_eq(vari="is")
ecrire_eq(vari="hd", ess="a")
}
}
