% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtrer_place.R
\name{Filtrer_place}
\alias{Filtrer_place}
\title{Filtering plots and preparing covariates for a simulation with Natura 3.0 growth simulator}
\usage{
Filtrer_place(fichier)
}
\arguments{
\item{fichier}{Dataframe of trees or of plots with columns: type_eco, sdom_bio, origine, iqs_pot_epn, iqs_pot_epb, iqs_pot_pig, iqs_pot_sab, iqs_pot_tho, iqs_pot_pib, iqs_pot_bop, iqs_pot_pex}
}
\value{
Dataframe filtered and with all necessary plots covariates.
}
\description{
Filter plots and prepare covariates for all the models of Natura 3.0 growth simulator.
}
\details{
The plots are filtered for type_eco, sdom_bio and origine values. Variables veg_pot and milieu are extracted from type_eco, binary variables are created for each category of sdom_bio, veg_pot, milieu and origine.
Variables containing potential iqs are renamed. The variable annee=0 is created.
}
\examples{
\dontrun{
plots <- Filtrer_place(fichier=liste_plot)
}
}
