% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lecture_fichier_arbres.R
\name{Lecture_arbres}
\alias{Lecture_arbres}
\title{Lire le fichier des arbres à simuler avec Natura et valider le nom des colonnes}
\usage{
Lecture_arbres(file, ht, vol, iqs, climat, sol)
}
\arguments{
\item{file}{Nom du fichier des arbres à lire (table, Excel ou csv)}

\item{ht}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si la hauteur des arbres doit être estimée, seulement pour un fichier \code{file_arbre}
  \item \code{FALSE}: si la hauteur des arbres est fournie dans \code{file_arbre}, ignoré si \code{file_compile}. Le nom de la colonne doit être hauteur_pred (en mètres)
  }}

\item{vol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si le volume des arbres doit être estimé, seulement pour un fichier \code{file_arbre}
  \item \code{FALSE}: si le volume des arbres est fourni dans \code{file_arbre}, ignoré si \code{file_compile}. Le nom de la colonne doit être vol_dm3 (en dm3)
  }}

\item{iqs}{Booléen
\itemize{
   \item \code{TRUE} (par défaut): si les IQS potentils doivent être extraits des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
   \item \code{FALSE}: si les IQS potentiels (m, hauteur dominante à 50 ans) sont fournis dans \code{file_arbre} ou \code{in file_compile}.
 }}

\item{climat}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables climatiques doivent être extraites des cartes. Les colonnes latitude, longitude et l'année de mesure de la placette doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \item \code{FALSE}: si les variables climatiques sont fournies dans \code{file_arbre} ou \code{file_compile}.
}}

\item{sol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables sur les propriétés des sols sont à extraires des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \item \code{FALSE}: si les variables sur les propriétés des sols sont fournies dans \code{file_arbre} ou dans \code{file_compile}.
}}
}
\value{
Table dont les arbres ont été filtrés ou un message d'erreur s'il y a une erreur dans le nom des colonnes.
}
\description{
Lire le fichier des arbres à simuler avec Natura et valider le nom des colonnes. Filtrer les arbres selon leur état et leur dhp.
}
