% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lecture_fichier_compile.R
\name{Lecture_compile}
\alias{Lecture_compile}
\title{Function reading the compiled plot file to be simulated with Natura 3.0}
\usage{
Lecture_compile(file, iqs, climat, sol)
}
\arguments{
\item{file}{Name of the file to read (dataframe, Excel or csv file)}

\item{iqs}{A boolean
\itemize{
   \item \code{TRUE} (default): if potential site index are to be extrated from maps. Columns latitude and longitude must be provided in \code{file_arbre} or in \code{file_compile}.
   \itemize{
      \item latitude, longitude: coordinates of plots, in decimal degrees
   }
   \item \code{FALSE}: if potential site index (m) are provided in \code{file_arbre} or \code{in file_compile}. Column names must be:
   \itemize{
     \item iqs_pot_sab, iqs_pot_epn, iqs_pot_epb, iqs_pot_pig, iqs_pot_tho, iqs_pot_pib, iqs_pot_bop, iqs_pot_pex.
   }
 }}

\item{climat}{A boolean
\itemize{
   \item \code{TRUE} (default): if climate variables are to be extrated from maps. Columns latitude, longitude, an_mes must be provided in \code{file_arbre} or in \code{file_compile}.
   \itemize{
     \item latitude, longitude: coordinates of plots, in decimal degrees
     \item an_mes: measurement year of the plot
   }
  \item \code{FALSE}: if climate variables are in \code{file_arbre} or in \code{file_compile}. Column names must be:
  \itemize{
    \item p_tot: 1980-2010 mean total precipitations (mm)
    \item t_ma: 1980-2010 mean annual temprature (Celcius)
    \item prec_gs: last 10-years mean growing season precipitations (mm)
    \item temp_gs: last 10-years mean growing season temperature (Celcius)
  }
}}

\item{sol}{A boolean
\itemize{
  \item \code{TRUE} (default): if soil variables are to be extrated from maps. Columns latitude and longitude must be provided in \code{file_arbre} or in \code{file_compile}.
  \itemize{
    \item latitude, longitude: coordinates of plots, in decimal degrees
    }
  \item \code{FALSE}: if SIIGSOL soil variables are in \code{file_arbre} or in \code{file_compile}. Soils properties at depth 0-5 cm. Column names must be:
  \itemize{
    \item ph: pH
    \item clay: fraction of clay
    \item silt: fraction of silt
    \item sand: fraction of sand
    \item cec: capacité d'échange cationique
    \item oc: matière organique
  }
}}
}
\value{
Dataframe at the plot scale ready to be simulate or an error message
}
\description{
Read a compiled plot file to be simulated with Natura 3.0. check the names of the variables, create total and  percentage variables, rename variables.
}
