% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lecture_fichier_etudes.R
\name{Lecture_etudes}
\alias{Lecture_etudes}
\title{Lire le fichier des arbres-études et valider le nom des colonnes}
\usage{
Lecture_etudes(file)
}
\arguments{
\item{file}{Nom du fichier à lire (table, Excel ou csv)}
}
\value{
Table arbres-études ou un message d'erreur s'il y a une erreur dans le nom des colonnes.
}
\description{
Lire le fichier des arbres-études et valider le nom des colonnes.
}
