% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Natura_oneStep.R
\name{Natura_oneStep}
\alias{Natura_oneStep}
\title{Simulation d'un pas de simulation (période) avec le modèle Natura}
\usage{
Natura_oneStep(
  data,
  param,
  data_info,
  long_int,
  pas,
  dec_perturb = 0,
  dec_tbe1 = 0,
  tbe1 = 0,
  dec_tbe2 = 0,
  tbe2 = 0
)
}
\arguments{
\item{data}{Table contenant les placettes et leurs caractéristiques dendrométriques (Nxxx, STxxx, Vxxx, HD, IS). Une ligne par placette. Ne doit pas contenir les covariables ni les paramètres.}

\item{param}{Table contenant les paramètres du modèle associés à chacune des placettes. Une ligne par placette.}

\item{data_info}{Table contenant les covariables fixes dans le temps pour chacune des placettes. Une ligne par placette.}

\item{long_int}{Nombre d'années d'un pas de simulation.}

\item{pas}{Le numéro du pas de simulation à simuler}

\item{dec_perturb}{Optionnel. Le numéro de la décennie dans laquelle on simule la présence d'une perturbation partielle autre qu'une épidémie de TBE (chablis partiel, verglas pertiel, dépérissement partiel, brulis partiel).}

\item{dec_tbe1}{Optionnel. Le numéro de la décennie dans laquelle on simule une défoliation due à la TBE. \code{tbe1} doit alors est spécifié.}

\item{tbe1}{Optionnel. Indice de réduction de croissance due à la TBE durant la décennie \code{dec_tbe1}, un entier 1 à 5. \code{dec_tbe1} doit alors est spécifié.
\itemize{
  \item 1 = 1 année de défoliation sévères
  \item 2 = 4 années de défoliations sévères
  \item 3 = 6 années de défoliations sévères
  \item 4 = 8 années de défoliations sévères
  \item 5= 10 années de défoliations sévères.
  }}

\item{dec_tbe2}{Optionnel. Le numéro de la 2e décennie dans laquelle on simule une défoliation due à la TBE. \code{tbe2} doit alors est spécifié.}

\item{tbe2}{Optionnel. Indice de réduction de croissance due à la TBE durant la décennie \code{dec_tbe2}, un entier 1 à 5. Voir \code{tbe1} pour définition des niveaux.}
}
\value{
Table contenant les caractéristiques dendrométriques prédites par le modèle pour le pas de simulation demandé. Une ligne par placette.
}
\description{
Simulation d'un pas de simulation (période) avec le modèle Natura, habituellement de 10 ans.
}
