% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametres_stochastiques.R
\name{param_evol_n_st_v_stoch}
\alias{param_evol_n_st_v_stoch}
\title{Generating n, st and v growth models parameter values of Natura 3.0}
\usage{
param_evol_n_st_v_stoch(liste_place, nb_iter, mode_simul, horizon, liste_ess)
}
\arguments{
\item{liste_place}{Dataframe with plot id (id_pe)}

\item{nb_iter}{Number of iteration if stochastic mode, ignored if \code{mode_simul="DET"}}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{horizon}{Number of 10-year time steps to simulate, 1 à 15}

\item{liste_ess}{Vector of species group code name of Natura 3.0 models}
}
\value{
A list of lists, one list per species group for which there is n-st-v growth models, and for each species group, a list with 3 elements:
\enumerate{
  \item essence: string with species group code name
  \item effet_fixe: dataframe of fixed effect parameter values (one line per time step per iteration)
  \item erreur_residuelle: dataframe of residual error values (one line per plot per time step per iteration)
}
}
\description{
Generate parameters of n, st and v growth models of Natura 3.0, deterministic or stochastic, for each plot for each time step and for all iterations
}
\details{
The n, st and v growth models of Natura 3.0 form a system of linear equations calibrated for each species groups. There is no random effect.
The plot residual errors are corralated among equations. There is a covariance matrix of fixed effects.
}
