% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametres_stochastiques.R
\name{param_hdom0_ess_stoch}
\alias{param_hdom0_ess_stoch}
\title{Génère les paramètres du modèle pour estimer la hauteur des arbres par essence à la step 0 pour le calcul de la hauteur dominante.}
\usage{
param_hdom0_ess_stoch(
  liste_arbre,
  mode_simul = "DET",
  nb_iter = 1,
  seed_value = NULL
)
}
\arguments{
\item{liste_arbre}{Table contenant l'identifiant de l'arbre (no_arbre) et l'identifiant de la placette (id_pe) pour lesquels la hauteur doit être estimée.}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{nb_iter}{Si mode stochastique, le nombre d'itérations à effectuer (> 1), ignoré si \code{mode_simul="DET"}}

\item{seed_value}{Optionnel. Valeur de départ pour la génération de nombres aléatoires. Utilisé lors des tests pour la reproductibilité des résultats.}
}
\value{
Une liste de listes. Une liste par essence pour lesquelles il y a un modèle de hauteur, et pour chaque essence, une liste avec ces 3 éléments:
\enumerate{
  \item essence: une chaine de caractère contenant le code de l'essence
  \item effet_fixe: table contenant les paramètres des effets fixes, une ligne par itération
  \item erreur_arbre: table contenant les valeurs des erreurs résiduelles et les valeurs des effets aléatoires d'arbres, une ligne par arbre par itération
}
}
\description{
Génère les paramètres du modèle pour estimer la hauteur des arbres par essence à la step 0 pour le calcul de la hauteur dominante, déterministe ou stochastique, pour chacun des arbres de chaque placette pour toutes les itérations.
}
\details{
Le modèle est un modèle non-linéaire mixte calibré par essence, avec un effet aléatoire d'arbre. Il n'y a pas de matrice de corrélation sur les résidus.
Il n'y a pas de matrice de covariances des effets fixes car il n'y a qu'un seul paramètre.
}
