% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametres_stochastiques.R
\name{param_is_evol_stoch}
\alias{param_is_evol_stoch}
\title{Generating shannon index growth model parameter values of Natura 3.0}
\usage{
param_is_evol_stoch(liste_place, nb_iter, mode_simul, horizon)
}
\arguments{
\item{liste_place}{Dataframe with plot id (id_pe)}

\item{nb_iter}{Number of iteration if stochastic mode, ignored if \code{mode_simul="DET"}}

\item{mode_simul}{Simulation mode: \code{"DET"} : deterministic or \code{"STO"} : stochastic}

\item{horizon}{Number of 10-year time steps to simulate, 1 à 15}
}
\value{
A list of lists, one list per species group for which there is a shannon index growth model, and for each species group, a list with 4 elements:
\enumerate{
  \item essence: string with species group code name
  \item effet_fixe: dataframe of fixed effect parameter values (one line per time step per iteration per species)
  \item random_placette: dataframe of random plot effect values (one line per plot per time step per iteration per species)
  \item erreur_residuelle: dataframe of residual error values (one line per plot per time step per iteration per species)
}
}
\description{
Generate parameters for shannon index growth model of Natura 3.0, deterministic or stochastic, for each plot for each time step and for all iterations
}
\details{
The shannon index growth model is a linear mixed model calibrated for each dominant species. There is a plot random effect
and the plot residual errors are not corralated. There is a covariance matrix of fixed effects.
}
