% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_fichier_arbres.R
\name{Prep_arbres}
\alias{Prep_arbres}
\title{Compilation des placettes à la step 0}
\usage{
Prep_arbres(
  fic_arbre,
  ht,
  vol,
  nb_iter = 1,
  nb_step = 1,
  mode_simul = "DET",
  seed_value = NULL,
  dt = dt
)
}
\arguments{
\item{fic_arbre}{Table avec une ligne par arbre (ou par classe de dhp) pour chaque placette, avec les colonnes : id_pe, sdom_bio, altitude, veg_pot, dummy vp, milieu, dummy mp, type_eco, p_tot, t_ma, prec_gs, temp_gs, dummy origine, temps, iqs_xxx, ph, cec, oc, sand, clay, no_arbre, essence, dhpcm, nb_tige, tige_ha, etat, (hauteur_pred, vol_dm3 optionnels)}

\item{ht}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si la hauteur des arbres doit être estimée, seulement pour un fichier \code{file_arbre}
  \item \code{FALSE}: si la hauteur des arbres est fournie dans \code{file_arbre}, ignoré si \code{file_compile}. Le nom de la colonne doit être hauteur_pred (en mètres)
  }}

\item{vol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si le volume des arbres doit être estimé, seulement pour un fichier \code{file_arbre}
  \item \code{FALSE}: si le volume des arbres est fourni dans \code{file_arbre}, ignoré si \code{file_compile}. Le nom de la colonne doit être vol_dm3 (en dm3)
  }}

\item{nb_iter}{Si mode stochastique, le nombre d'itérations à effectuer (> 1), ignoré si \code{mode_simul="DET"}}

\item{nb_step}{Le nombre d'années pour lesquelles on veut estimer la hauteur pour un même arbre (par défaut 1).}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{seed_value}{Optionnel. Valeur de départ pour la génération de nombres aléatoires. Utilisé lors des tests pour la reproductibilité des résultats.}

\item{dt}{La durée de l'intervalle de temps entre deux mesures d'un même arbre si \code{nb_step>1} (par défaut 10).}
}
\value{
Une liste de 2 éléments: table contenant les arbres avec leur hauteur et volume (une ligne par arbre) et une table des caractérisatiques des placettes (une ligne par placette)
}
\description{
Compilation des placettes à la step 0: estime la hauteur et le volume des arbres si nécessaire, calcule l'indice de Shannon et calcule les caractéristiques dendrométriques (N, ST, V) par groupe d'essences.
}
