% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_fichier_etude.R
\name{Prep_etude}
\alias{Prep_etude}
\title{Calcule la hauteur dominante à la step 0}
\usage{
Prep_etude(
  fic_etude,
  fic_arbre,
  nb_iter = 1,
  mode_simul = "DET",
  seed_value = NULL
)
}
\arguments{
\item{fic_etude}{Table des arbres-études}

\item{fic_arbre}{Tables des arbres averc leur hauteur estimée, avec les itérations si stochastiques}

\item{nb_iter}{Si mode stochastique, le nombre d'itérations à effectuer (> 1), ignoré si \code{mode_simul="DET"}}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{seed_value}{Optionnel. Valeur de départ pour la génération de nombres aléatoires. Utilisé lors des tests pour la reproductibilité des résultats.}
}
\value{
Table contenant la hauteur dominante de chaque placette, une ligne par placette/iter.
}
\description{
Calcule la hauteur dominante des placettes à la step 0 à partir des arbres-études et du fichier des arbres.
}
