% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_parametre.R
\name{Prep_parametre_iter}
\alias{Prep_parametre_iter}
\title{Select the growth model parameters associated with an iteration for a list of plots for a simulation with Natura 3.0}
\usage{
Prep_parametre_iter(iteration, param_ishd_evol, param_n_st_v, liste_ess)
}
\arguments{
\item{iteration}{Iteration number}

\item{param_n_st_v}{Object with n-st-v growth models parameter values for all time steps and all iterations}

\item{liste_ess}{Vector with Natura group species code names}
}
\value{
Dataframe of one observation with all the growth models parameter values for an iteration
}
\description{
Select the growth models parameters associated with an iteration.
}
