% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_parametre.R
\name{Prep_parametre_pas}
\alias{Prep_parametre_pas}
\title{Sélectionne les paramètres des modèles associés à un pas de simulation, pour le cas où un fichier à l'échelle de la placette est fourni}
\usage{
Prep_parametre_pas(
  data,
  data_info,
  iteration,
  pas,
  mode_simul,
  param_ishd_evol,
  param_n_st_v,
  liste_ess
)
}
\arguments{
\item{data}{Table contenant seulement l'identifiant des placettes (id_pe) et les variables contenant le pourcentage de surface terrière de chaque groupe d'essences (pct_epn1, pct_epx1, etc),  une ligne par placette}

\item{data_info}{Table contenant seulement les colonnes associées aux caractéristiques des placettes qui ne changent pas dans le temps}

\item{iteration}{Numéro de l'itération}

\item{pas}{Numéro du pas de simulation}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{param_ishd_evol}{Objet contenant les paramètres du modèle de l'indice de Shannon et du modèle de hauteur dominante, pour toutes les itérations et pas de simulation}

\item{param_n_st_v}{Objet contenant les paramètres des modèles de Nxxx, STxxx et Vxxx, pour toutes les itération et pas de smulation}

\item{liste_ess}{Vecteur contenant les codes des groupes d'essences de Natura}
}
\value{
Table contenant toutes les colonnes nécessaires pour effectuer un pas de simulation pour un pas de simulation.
}
\description{
Sélectionne les paramètres des modèles associés à un pas de simulation, pour le cas où un fichier à l'échelle de la placette est fourni en entrée.
Ces paramètres sont ceux qui changent à chauqe pas de simulation pour une itération donnée, soit les erreurs résiduelles.
Ajoute les paramètres au fichier compilé à la placette. Détermine le groupe d'essences dominant et sélectionne l'IQS associé.
}
