% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_columns.R
\name{remove_columns}
\alias{remove_columns}
\title{Cette fonction supprime les colonnes spécifiées d'un data frame si elles existent.}
\usage{
remove_columns(data, columns_to_remove)
}
\arguments{
\item{data}{Un data frame à partir duquel les colonnes seront supprimées.}

\item{columns_to_remove}{Un vecteur de caractères contenant les noms des colonnes à vérifier et à supprimer si elles existent dans le data frame.}
}
\value{
Un data frame avec les colonnes spécifiées supprimées si elles existent.
}
\description{
Cette fonction supprime les colonnes spécifiées d'un data frame si elles existent.
}
