% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateurNatura.R
\name{SimulNatura}
\alias{SimulNatura}
\title{Fonction principale pour une simulation de la croissance d'un peuplement avec le simulateur Natura 3.0}
\usage{
SimulNatura(
  file_arbre,
  file_etude,
  file_compile,
  file_export,
  horizon,
  mode_simul = "DET",
  nb_iter = 1,
  iqs = TRUE,
  climat = TRUE,
  sol = TRUE,
  ht = TRUE,
  vol = TRUE,
  dec_perturb = 0,
  dec_tbe1 = 0,
  tbe1 = 0,
  dec_tbe2 = 0,
  tbe2 = 0,
  seed_value = NULL
)
}
\arguments{
\item{file_arbre}{Nom du fichier contenant les informations sur les arbres et les placettes (table, xlsx ou csv). Le fichier doit contenir une ligne par arbre ou par classe de dhp/essence.
Si ce paramètre n'est pas utilisé, \code{file_compile} doit l'être. Le nom des colonnes doit être:
 \itemize{
   \item id_pe: identifiant unique de la placette
   \item origine: code d'origine de la placette (ES: épidémie sévère, BR: brulis, CT: coupe totale)
   \item temps_origine: temps depuis l'origine (ans)
   \item sdom_bio: code du sous-domaine bioclimatique, en majuscule (ex: 2E, 4O)
   \item altitude: altitude (m)
   \item type_eco: code du type écologique (ex: MS22)
   \item essence: code de l'essences de l'arbre (ex: SAB, EPN, BOP)
   \item dhpcm: dhp (cm) de l'arbre ou classe de dhp (seuls les arbres de plus de 9 cm sont retenus)
   \item nb_tige: nombre d'arbres de l'essence et de la classe de dhp, à l'ha
   \item etat: code d'état de l'arbre, seuls les vivants sont retenus, codes 10, 12, 30, 32, 40, 42, 50, 52
 }}

\item{file_etude}{Nom du fichier contenant les arbres-études (table, xlsx ou csv). Le fichier doit contenir une ligne par arbre.
Si ce paramètre n'est pas utilisé, \code{file_compile} doit l'être. Le nom des colonnes doit être:
\itemize{
  \item id_pe: identifiant unique de la placette
  \item étage: code d'étage de l'arbre: C, D, O, I, V, seuls les C et D sont retenus
  \item essence: code de l'essences de l'arbre-étude (ex: SAB, EPN, BOP)
  \item dhpcm: dhp de l'arbre-étude (cm)
  \item hauteur: hauteur totale de l'arbre-étude (m)
}}

\item{file_compile}{Optionnel. Nom du fichier contenant les caractéristiques dendrométriques des placettes (table, xlsx ou csv). Une ligne par placette.
 Si ce paramètre n'est pas utilisé, \code{file_arbre} et \code{file_etude} doivent l'être. La composition de chacun des groupes d'essences est décrite dans la section Details.
 Le nom des colonnes doit être:
 \itemize{
   \item id_pe: identifiant unique de la placette
   \item origine: code d'origine de la placette (ES: épidémie sévère, BR: brulis, CT: coupe totale)
   \item temps_origine: temps depuis l'origine (ans)
   \item sdom_bio: code du sous-domaine bioclimatique, en majuscule (ex: 2E, 4O)
   \item altitude: altitude (m)
   \item type_eco: code du type écologique (ex: MS22)
   \item is: indice de structure diamètrale de Shannon (valeur entre 0 et 1; 0 : toutes les tiges sont dans la même classe de dhp; 1 : les tiges sont également distribuées dans les classes de dhp)
   \item hd: hauteur dominante, hauteur moyenne des 100 plus gros arbres à l'ha (m)
   \item nxxx : nombre d'arbres de plus de 9 cm du groupe d'essences xxx (/ha), mettre 0 si absent. xxx : chacun des groupes d'essences (bop, peu, ft, ri, rt, epx, epn, sab)
   \item stxxx: surface terrière marchande du groupe d'essences xxx (m2/ha), mettre 0 si absent. xxx : chacun des groupes d'essences (bop, peu, ft, ri, rt, epx, epn, sab)
   \item vxxx: volume marchand brut du groupe d'essences xxx (m3/ha), mettre 0 si absent. Si aucun volume, mettre les 8 groupes à 0 et l'évolution du volume ne sera pas effectué. xxx : chacun des groupes d'essences (bop, peu, ft, ri, rt, epx, epn, sab)
}}

\item{file_export}{Nom et emplacement du fichier à exporter, contenant les résultats de la simulation (format csv), optionnel.}

\item{horizon}{Nombre de décennies à simuler, un chiffre de 1 à 15.}

\item{mode_simul}{Mode de simulation: \code{"DET"} : déterministe ou \code{"STO"} : stochastique}

\item{nb_iter}{Si mode stochastique, le nombre d'itérations à effectuer (> 1), ignoré si \code{mode_simul="DET"}}

\item{iqs}{Booléen
\itemize{
   \item \code{TRUE} (par défaut): si les indices de qualité de station potentielle doivent être extraits des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
   \itemize{
      \item latitude, longitude: coordonnées des placettes, en degré décimal
   }
   \item \code{FALSE}: si les indices de qualité de station potentielle (m) sont fournis dans \code{file_arbre} ou \code{in file_compile}. Le nom des 8 colonnes doit être:
   \itemize{
     \item iqs_pot_sab, iqs_pot_epn, iqs_pot_epb, iqs_pot_pig, iqs_pot_tho, iqs_pot_pib, iqs_pot_bop, iqs_pot_pex.
   }
 }}

\item{climat}{Booléen
\itemize{
   \item \code{TRUE} (par défaut): si les variables climatiques doivent être extraites des cartes. Les colonnes latitude, longitude et l'année de mesure de la placette doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
   \itemize{
     \item latitude, longitude: coordonnées des placettes, en degré décimal
     \item an_mes: année de mesure de la placette
   }
  \item \code{FALSE}: si les variables climatiques sont fournies dans \code{file_arbre} ou \code{file_compile}. Le nom des 4 colonnes doit être:
  \itemize{
    \item p_tot: précipitations totales annuelles, moyenne sur la période 1980-2010 (mm)
    \item t_ma: température moyenne annuelle sur la période 1980-2010 (Celcius)
    \item prec_gs: précipitations durant la saison de croissance, moyenne sur les 10 dernières années avant le mesurage de la placette (mm)
    \item temp_gs: température moyenne durant la saison de croissance, moyenne sur les 10 dernières années avant le mesurage de la placette (Celcius)
  }
}}

\item{sol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si les variables sur les propriétés des sols sont à extraires des cartes. Les colonnes latitude et longitude doivent alors être fournies dans \code{file_arbre} ou dans \code{file_compile}.
  \itemize{
    \item latitude, longitude: ccoordonnées des placettes, en degré décimal
    }
  \item \code{FALSE}: si les variables sur les propriétés des sols (SIIGSOL) sont fournies dans \code{file_arbre} ou dans \code{file_compile}. Propriétés des sols dans la profondeurt 0-5 cm. Le nom des 6 colonnes doit être:
  \itemize{
    \item ph: pH
    \item clay: proportion d'argile (valeur de 0 à 100)
    \item sand: proportion de sable (valeur de 0 à 100)
    \item cec: capacité d'échange cationique (méq/100g)
    \item oc: proportion de matière organique (valeur de 0 à 100)
  }
}}

\item{ht}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si la hauteur des arbres doit être estimée, si \code{file_arbre} est fourni
  \item \code{FALSE}: si la hauteur des arbres est fournie dans \code{file_arbre}, ignoré si \code{file_compile} est fourni. Le nom de la colonne doit être hauteur_pred (en mètres)
  }}

\item{vol}{Booléen
\itemize{
  \item \code{TRUE} (par défaut): si le volume des arbres doit être estimé, si \code{file_arbre} est fourni
  \item \code{FALSE}: si le volume des arbres est fourni dans \code{file_arbre}, ignoré si \code{file_compile} est fourni. Le nom de la colonne doit être vol_dm3 (en dm3)
  }}

\item{dec_perturb}{Optionnel. Le numéro de la décennie dans laquelle on simule la présence d'une perturbation partielle autre qu'une épidémie de TBE (chablis partiel, verglas pertiel, dépérissement partiel, brulis partiel).}

\item{dec_tbe1}{Optionnel. Le numéro de la décennie dans laquelle on simule une défoliation due à la TBE. \code{tbe1} doit alors est spécifié.}

\item{tbe1}{Optionnel. Indice de réduction de croissance due à la TBE durant la décennie \code{dec_tbe1}, un entier 1 à 5. \code{dec_tbe1} doit alors est spécifié.
\itemize{
  \item 1 = 1 année de défoliation sévères
  \item 2 = 4 années de défoliations sévères
  \item 3 = 6 années de défoliations sévères
  \item 4 = 8 années de défoliations sévères
  \item 5= 10 années de défoliations sévères.
  }}

\item{dec_tbe2}{Optionnel. Le numéro de la 2e décennie dans laquelle on simule une défoliation due à la TBE. \code{tbe2} doit alors est spécifié.}

\item{tbe2}{Optionnel. Indice de réduction de croissance due à la TBE durant la décennie \code{dec_tbe2}, un entier 1 à 5. Voir \code{tbe1} pour définition des niveaux.}

\item{seed_value}{Optionnel. Valeur de départ pour la génération de nombres aléatoires. Utilisé lors des tests pour la reproductibilité des résultats.}
}
\value{
Une table contenant les résultats de la simulation, une ligne par placette/décennie/itération:
 \itemize{
   \item id_pe: identifiant unique de la placette
   \item iter: Numéro de l'itération si le mode stochastique a été utilisé, la colonne n'est pas présente si mode déterministe
   \item sdom_bio: code du sous-domaine bioclimatique, en majuscule
   \item type_eco: code du type écologique
   \item origine: code d'origine de la placette (ES: épidémie sévère, BR: brulis, CT: coupe totale)
   \item annee: nombre d'années depuis le début de la simulation (ans)
   \item temps: temps depuis l'origine (ans)
   \item tbe: indice de réduction de croissance due à la TBE durant la décennie
   \item pert: présence d'une perturbation partielle autre qu'une épidémie de TBE dans la décennie (1=présence, 0=absence)
   \item hd: hauteur dominante, hauteur moyenne des 100 plus gros arbres à l'ha (m)
   \item is: indice de structure diamètrale de Shannon
   \item stxxx: surface terrière marchande du groupe d'essences xxx (m2/ha) (bop, peu, ft, ri, rt, epx, epn, sab)
   \item sttot: surface terrière marchande de l'ensemble des essences (m2/ha)
   \item nxxx : nombre d'arbres de plus de 9 cm du groupe d'essences xxx (tiges/ha) (bop, peu, ft, ri, rt, epx, epn, sab)
   \item ntot : nombre d'arbres de plus de 9 cm de l'ensemble des essences (tiges/ha)
   \item vxxx: volume marchand brut du groupe d'essences xxx (m3/ha) (bop, peu, ft, ri, rt, epx, epn, sab)
   \item vtot: volume marchand brut du groupe de l'ensemble des essences (m3/ha)
   \item pct_xxx: pourcentage de la surface terrière du groupe d'essences xxx (pourcentage) (bop, peu, ft, ri, rt, epx, epn, sab)
   \item dqxxx: diamètre quadratique moyen du groupe d'essences xxx (cm) (bop, peu, ft, ri, rt, epx, epn, sab)
   \item dqtot: diamètre quadratique moyen de l'ensemble des essences (cm)
}
}
\description{
\code{SimulNatura()} est la fonction principale pour exécuter une simulation de la croissance d'un peuplement avec le simulateur Natura 3.0 (version qui utilise l'origine du peuplement et le temps depuis l'origine).
}
\details{
Le simulateur fonctionne pour des peuplements situés sur les végétations potentielles suivantes:
\itemize{
   \item ME1, MS2, MS4, MS6
   \item RB1, RB5,
   \item RP1
   \item RE1, RE2, RE3, RE4
   \item RS1, RS2, RS3, RS4, RS5, RS7
   }
Les groupes d'essences du simulateur sont composés des essences suivantes:
\itemize{
   \item bop: bouleau à papier
   \item peu: peuplier faux-tremble, peuplier baumier, peuplier à grandes dents
   \item ft: feuillus tolérants: érable rouge, bouleau jaune, érable à sucre, frêne noir
   \item epn: épinette noire
   \item epx: épinette blanche, épinette rouge
   \item sab: sapin baumier
   \item ri: résineux intolérants: pin gris, mélèze laricin, pin rouge
   \item rt: résineux tolérants: thuya, pin blanc, pruche
   }
}
\examples{
\dontrun{
# Simulation déterministe sur 50 ans à partir d'un fichier à l'échelle de l'arbre, hauteur et volume des arbres doit être estimés et les variables d'IQS, de climat et de sol doivent être extraites
data_simul <- SimulNatura(file_arbre=fichier_arbres_sanscov, file_etude=fichier_arbres_etudes, horizon=5)

#' # Simulation déterministe sur 50 ans à partir d'un fichier à l'échelle de l'arbre, hauteur et volume des arbres doit être estimés, mais les variables d'IQS, de climat et de sol sont fournies dans le fichier d'entrée
data_simul <- SimulNatura(file_arbre=fichier_arbres_aveccov, file_etude=fichier_arbres_etudes, horizon=5, iqs=FALSE, sol=FALSE, climat=FALSE)

# Simulation stochastique sur 50 ans à partir d'un fichier à l'échelle de l'arbre, hauteur et volume des arbres doit être estimés, mais les variables d'IQS, de climat et de sol sont fournies dans le fichier d'entrée
data_simul <- SimulNatura(file_arbre=fichier_arbres_aveccov, file_etude=fichier_arbres_etudes, horizon=5, mode_simul='STO', nb_iter=30, iqs=FALSE, sol=FALSE, climat=FALSE)

# Simulation déterministe sur 50 ans à partir d'un fichier à l'échelle de la placette, les variables d'IQS, de climat et de sol sont fournies dans le fichier d'entrée
data_simul <- SimulNatura(file_compile=fichier_compile_aveccov, horizon=5, iqs=FALSE, sol=FALSE, climat=FALSE)
}
}
