% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_columns.R
\name{valide_placette_etudes}
\alias{valide_placette_etudes}
\title{Valide la correspondance des placettes entre le fichier des arbres et celui des arbres-études}
\usage{
valide_placette_etudes(data_arbre, data_etude)
}
\arguments{
\item{data_arbre}{Nom du fichier contenant les informations sur les arbres et les placettes}

\item{data_etude}{Nom du fichier contenant les arbres-études}
}
\value{
Une liste de messages d'erreurs. Si aucune erreur n'est trouvée, une liste vide est retournée.
}
\description{
Cette fonction vérifie que toutes les placettes présentes dans le fichier des arbres sont également présentes dans le fichier des arbres-études
}
\examples{

data_arbre <- data.frame(Placette = c("TEM23APC5001", "TEM23APC5002", "TEM23APC5003"))
data_etude <- data.frame(Placette = c("TEM23APC5001", "TEM23APC5002"))
valide_placette_etudes(data_arbre, data_etude)
# Les erreurs seront retournées comme ceci
# 'Aucune des placettes suivantes ne sont valide dans le fichier des arbres-études : TEM23APC5003'
}
